services/ai/pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.129.0",
    "chromadb>=1.1.0",
    "httpx>=0.28.1",
    "openai>=1.108.0",
    "pyjwt[crypto]>=2.10.1",
]
